requirements.txt
anyio==3.6.2
argcomplete==2.0.0
attrs==22.2.0
-e git+https://github.com/RepoAnalyzer/awesome-list-miner.git@068780a6276c8a2f8e67bc24f68dabfb7c772ced#egg=awesome_list_miner
-e git+https://github.com/RepoAnalyzer/awesome-list-miner.git@068780a6276c8a2f8e67bc24f68dabfb7c772ced#egg=awesome_list_parser
beautifulsoup4==4.11.2
black==22.12.0
bleach==6.0.0
bs4==0.0.1
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commitizen==2.41.0
decli==0.5.2
Deprecated==1.2.13
distlib==0.3.6
docutils==0.19
exceptiongroup==1.1.0
filelock==3.9.0
flake8==5.0.4
h11==0.14.0
httptools==0.5.0
identify==2.5.17
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
isort==5.12.0
jaraco.classes==3.2.3
Jinja2==3.1.2
jsii==1.74.0
keyring==23.13.1
Markdown==3.4.1
MarkupSafe==2.1.2
mccabe==0.7.0
more-itertools==9.0.0
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.0
pathspec==0.11.0
pkginfo==1.9.6
platformdirs==3.0.0
pluggy==1.0.0
pre-commit==2.21.0
projen==0.65.13
prompt-toolkit==3.0.36
publication==0.0.3
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.4
pyflakes==2.5.0
PyGithub==1.57
Pygments==2.14.0
PyJWT==2.6.0
PyNaCl==1.5.0
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pywin32-ctypes==0.2.0
PyYAML==6.0
questionary==1.10.0
readme-renderer==37.3
-e git+https://github.com/RepoAnalyzer/awesome-list-miner.git@068780a6276c8a2f8e67bc24f68dabfb7c772ced#egg=repo_analyzer_awesome_list_parser
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==2.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.24.0
termcolor==2.2.0
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
twine==3.3.0
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
virtualenv==20.19.0
watchfiles==0.18.1
wcwidth==0.2.6
webencodings==0.5.1
websockets==10.4
wrapt==1.14.1
zipp==3.12.1
